Fixed code not compiling on Ubuntu 20.04 / ROS Noetic
ROS Noetic is bundled with OpenCV 4, meaning that some old C compatible constants are not available anymore.
They were replaced by their C++ equivalent, which were also part of OpenCV 3, meaning that nothing should break on ROS Kinetic / Melodic.
An other modification also had to be done, to compile a part of the code that was restricted to OpenCV 2 or OpenCV 3.